Add initial draft of the chatbot
Document the workflow of the chatbot.

Define the LLM chains for the nodes as well as for the conditional
edges. Use structured output to reduce hallucinations.

Use a strict flow with the following binary classifications:
* whether a response is needed
* whether more info is needed from the customer
* whether the draft reply is helpful
* whether human intervention is needed for an unhelpful draft reply
* whether an issue tracker item should be created.

Loop up to RETRIES times to draft a reply, making use of the CoT of the
helpfulness classification.

Use query rewriting and RAG with Chroma.